requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
Flask==0.12.2
gunicorn==19.7.1
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
prompt-toolkit==1.0.15
python-Levenshtein==0.12.0
requests==2.18.3
selenium==3.4.3
simplejson==3.11.1
six==1.10.0
urllib3==1.22
wcwidth==0.1.7
Werkzeug==0.12.2
